ai_agent/requirements.txt
#langchain==0.2.14
#langchain-core==0.2.36
#langchain-community==0.2.10
#langchain-ollama==0.1.0
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.3.3
ollama==0.3.0
